+++ frontend/lib/modules/settings/model/theme_parser/kitty_terminal_theme_parser.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:ui';
 
-import 'package:cliq/modules/settings/model/terminal_theme_parser/terminal_theme_parser.dart';
+import 'package:cliq/modules/settings/model/theme_parser/terminal_theme_parser.dart';
 import 'package:cliq/shared/data/database.dart';
 import 'package:cliq/shared/extensions/color.extension.dart';
 import 'package:drift/drift.dart';
