+++ frontend/lib/modules/settings/model/theme_parser/terminal_theme_parser.dart
