+++ frontend/test/settings_importer_test.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:cliq/modules/settings/model/import/settings_importer.dart';
+import 'package:cliq/modules/settings/model/settings_importer/settings_importer.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'util/test_utils.dart';
